timekit: a toolkit for time series

The timekit package combines a collection of coercion tools for time series analysis.

Details

The timekit package has two main benefits:

It simplifies the coercion process between time-based tibbles (tbl) and the major time series data types xts, zoo, zooreg, and ts

It maximizes time-based data retention during coercion to regularized time series

To learn more about timekit, start with the vignettes: browseVignettes(package = "timekit")