Augment the time series signature to the data

Augment the time series signature to the data

tk_augment_timeseries_signature(data)

Arguments

data
A time-based tibble or time-series object.

Value

Returns a tibble object describing the timeseries.

Details

tk_augment_timeseries_signature adds the time series signature features including numeric value, differences, year, month, day, day of week, day of month, day of year, hour, minute, second to the input data.

See also

tk_get_timeseries_signature(), tk_get_timeseries_summary()

Examples

library(tidyquant)
library(timekit)

FANG %>%
    filter(symbol == "FB") %>%
    tk_augment_timeseries_signature()

#> # A tibble: 1,008 × 30
#>    symbol       date  open  high   low close    volume adjusted  index.num
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>      <int>
#> 1      FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00 1357084800
#> 2      FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77 1357171200
#> 3      FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76 1357257600
#> 4      FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42 1357516800
#> 5      FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06 1357603200
#> 6      FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59 1357689600
#> 7      FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30 1357776000
#> 8      FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72 1357862400
#> 9      FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95 1358121600
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10 1358208000
#> # ... with 998 more rows, and 21 more variables: diff <int>, year <int>,
#> #   half <int>, quarter <int>, month <int>, month.xts <int>, month.lbl <ord>,
#> #   day <int>, hour <int>, minute <int>, second <int>, wday <int>,
#> #   wday.xts <int>, wday.lbl <ord>, mday <int>, yday <int>, week <int>,
#> #   week.iso <int>, week2 <int>, week3 <int>, week4 <int>